+++ composer.json
@@ -37,7 +37,8 @@
       "src/lib/api.php",
       "src/lib/helpers.php",
       "src/lib/list-table.php",
-      "src/lib/shell.php"
+      "src/lib/shell.php",
+      "src/modules/pages"
     ]
   },
   "scripts": {
